+++ bindings/nodejs/CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 1.1.4 - 2023-MM-DD
 
+### Added
+
+- `Utils:verifySemantic()`;
+
 ### Fixed
 
 - `StateMetadataOutput`'s constructor not setting the `stateMetadata` field;